requirements.txt
Flask
pymongo
dnspython
reportlab
gunicorn
requests
flask_cors
fpdf
apscheduler
cryptography
pillow
PyJWT
razorpay
setuptools
bs4
boto3
pyrebase4

